Remember user's name for next join
The app should now remember the user's name and autofills the name input
on the join screen.

I've added another `useEffect` instance to `AppContext.js` that sets the
username stored in `state.loginInfo.username` to `localStorage`. But it
will only do this if the value is truthy, i.e. not `""`.

The username is retrieved from `localStorage` in `PreviewScreen.jsx` and
stored in the variable `usernameFromStorage`. I then pass this value
into the `config` prop (via `convertLoginInfoToJoinConfig()`) of the
`Preview` component.

I've also added `USERNAME_KEY` to constants.js for setting/retrieving
the username from `localStorage`.